+++ Generator/Classes/Configuration/Configuration.php
@@ -15,5 +15,5 @@ class Configuration
     public $modelTypeTraitPathTemplate = '';
     public $modelTypePathTemplate = '';
     public $viewHelperTypePathTemplate = '';
-    public $webPageTypeProviderTemplate = '';
+    public $typesProviderTemplate = '';
 }