+++ Generator/generate.php
@@ -12,7 +12,7 @@ $configuration->schemaPath = __DIR__ . '/Schema/schema.jsonld';
 $configuration->modelTypeTraitPathTemplate = __DIR__ . '/../Classes/Model/TypeTrait/%sTrait.php';
 $configuration->modelTypePathTemplate = __DIR__ . '/../Classes/Model/Type/%s.php';
 $configuration->viewHelperTypePathTemplate = __DIR__ . '/../Classes/ViewHelpers/Type/%sViewHelper.php';
-$configuration->webPageTypeProviderTemplate = __DIR__ . '/../Classes/Provider/WebPageTypeProvider.php';
+$configuration->typesProviderTemplate = __DIR__ . '/../Classes/Provider/TypesProvider.php';
 
 $loader = new FilesystemLoader(__DIR__ . '/Templates');
 $twig = new Environment($loader, [